% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cusum.transform.R
\name{cusum.transform.missing}
\alias{cusum.transform.missing}
\title{MissCUSUM transformation of a matrix with missing entries}
\usage{
cusum.transform.missing(x)
}
\arguments{
\item{x}{a matrix with missing entries represented by NA}
}
\value{
MissCUSUM transformed matrix
}
\description{
MissCUSUM transformation of a matrix with missing entries
}
