% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cusum.transform.R
\name{cusum.univariate.missing}
\alias{cusum.univariate.missing}
\title{MissCUSUM transformation of a single vector with missing entries}
\usage{
cusum.univariate.missing(x)
}
\arguments{
\item{x}{a vector with missing entries represented by NA}
}
\value{
MissCUSUM transformed vector
}
\description{
MissCUSUM transformation of a single vector with missing entries
}
