\name{InspectChangepoint-package}
\alias{InspectChangepoint-package}
\alias{InspectChangepoint}
\docType{package}
\title{
\packageTitle{InspectChangepoint}
}
\description{
\packageDescription{InspectChangepoint}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{InspectChangepoint}
\packageIndices{InspectChangepoint}
}
\author{
\packageAuthor{InspectChangepoint}

Maintainer: \packageMaintainer{InspectChangepoint}
}
\references{
Wang, T. and Samworth, R. J. (2016) High-dimensional changepoint estimation via sparse projection. Arxiv preprint: arxiv1606.06246.
}
\keyword{ package }

