% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspect.R
\name{plot.inspect}
\alias{plot.inspect}
\title{Plot function for 'inspect' class objects}
\usage{
\method{plot}{inspect}(obj, ...)
}
\arguments{
\item{obj}{an 'inspect' class object}
}
\description{
Plot function for 'inspect' class objects
}
\seealso{
\code{\link{inspect}}
}
