% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspect.R
\name{print.inspect}
\alias{print.inspect}
\title{Print function for 'inspect' class objects}
\usage{
\method{print}{inspect}(obj, ...)
}
\arguments{
\item{obj}{an 'inspect' class object}
}
\description{
Print function for 'inspect' class objects
}
\seealso{
\code{\link{inspect}}
}
