% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary.R
\name{printPercentage}
\alias{printPercentage}
\title{Print percentage}
\usage{
printPercentage(ind, tot)
}
\arguments{
\item{ind}{a vector of for loop interator}

\item{tot}{a vector of for loop lengths}
}
\value{
on screen output of percentage
}
\description{
Print percentage
}
