% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparse.svd.R
\name{sparse.svd.missing}
\alias{sparse.svd.missing}
\title{Computing the sparse leading left singular vector of a matrix with missing entries}
\usage{
sparse.svd.missing(Z, lambda, max_iter = 1000, tol = 1e-10)
}
\arguments{
\item{Z}{Input matrix whose left leading singular vector is to be estimated.}

\item{lambda}{Regularisation parameter}

\item{max_iter}{maximum iteration}

\item{tol}{tolerance level for convergence}
}
\description{
Computing the sparse leading left singular vector of a matrix with missing entries
}
