% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspect.R
\name{summary.inspect}
\alias{summary.inspect}
\title{Summary function for 'inspect' class objects}
\usage{
\method{summary}{inspect}(obj, ...)
}
\arguments{
\item{obj}{an 'inspect' class object}
}
\description{
Summary function for 'inspect' class objects
}
\seealso{
\code{\link{inspect}}
}
