% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary.R
\name{vector.clip}
\alias{vector.clip}
\title{Clipping a vector from above and below}
\usage{
vector.clip(x, upper = Inf, lower = -upper)
}
\arguments{
\item{x}{a vector of real numbers}

\item{upper}{clip above this value}

\item{lower}{clip below this value}
}
\value{
the entrywise L_q norm of a vector or a matrix
}
\description{
Clipping vector or matrix x from above and below
}
