% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary.R
\name{vector.normalise}
\alias{vector.normalise}
\title{Normalise a vector}
\usage{
vector.normalise(v, q = 2, na.rm = FALSE)
}
\arguments{
\item{v}{a vector of real numbers}

\item{q}{a nonnegative real number or Inf}

\item{na.rm}{boolean, whether to remove NA before calculation}
}
\value{
normalised version of this vector
}
\description{
Normalise a vector
}
