% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspect.R
\name{plot.inspect}
\alias{plot.inspect}
\title{Plot function for 'inspect' class objects}
\usage{
\method{plot}{inspect}(x, ...)
}
\arguments{
\item{x}{an 'inspect' class object}

\item{...}{other arguments to be passed to methods are not used}
}
\description{
Plot function for 'inspect' class objects
}
\seealso{
\code{\link{inspect}}
}
