% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspect.R
\name{print.inspect}
\alias{print.inspect}
\title{Print function for 'inspect' class objects}
\usage{
\method{print}{inspect}(x, ...)
}
\arguments{
\item{x}{an 'inspect' class object}

\item{...}{other arguments to be passed to methods are not used}
}
\description{
Print function for 'inspect' class objects
}
\seealso{
\code{\link{inspect}}
}
