% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspect.R
\name{summary.inspect}
\alias{summary.inspect}
\title{Summary function for 'inspect' class objects}
\usage{
\method{summary}{inspect}(object, ...)
}
\arguments{
\item{object}{an 'inspect' class object}

\item{...}{other arguments to be passed to methods are not used}
}
\description{
Summary function for 'inspect' class objects
}
\seealso{
\code{\link{inspect}}
}
